refactor: add useTextInput hook, generalize the text input logic/UI
