+++ src/views/PublishView/index.ts
@@ -1,4 +1,3 @@
 import PublishView from "./PublishView";
 
 export default PublishView;
-export * from "./PublishView";
